+++ apps/strapi/src/api/modeles-de-courrier/routes/modeles-de-courrier.ts
@@ -1,0 +1,7 @@
+/**
+ * modeles-de-courrier router
+ */
+
+import { factories } from "@strapi/strapi";
+
+export default factories.createCoreRouter("api::modeles-de-courrier.modeles-de-courrier");