+++ apps/web/src/components/utils/Markdown.tsx
@@ -19,6 +19,13 @@ export const Markdown = ({ children, ...rest }: MarkdownProps) => {
           </div>
         ),
         img: ({ ...props }) => <img className="fr-fluid-img" {...props} />,
+        a: ({ node, children, ...props }) => {
+          if (props.href?.includes("http")) {
+            props.target = "_blank";
+            props.rel = "noreferrer";
+          }
+          return <a {...props}>{children}</a>;
+        },
       }}
       {...rest}
     >